pyproject.toml
[project]
name = "hermes"
version = "0.1.0"
description = "Alertmanager routing and distribution system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "watchdog>=4.0.0",
    "prometheus-client>=0.19.0",
    "requests>=2.31.0",
    "httpx>=0.26.0",
    "redis>=5.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"